Move GeoLocation override methods to a separate DevToolsEmulationExtensions class
